# ui/web/src/main/resources/public/swagger/index.yaml
swagger: '2.0'
info:
  description: 是一套轻量级权限管理系统，极低门槛，拿来即用。设计之初，就非常注重安全性，为企业系统保驾护航，让一切都变得如此简单。
  version: 3.0.0
  title: vmes-admin
basePath: /vmes-admin
schemes:
  - http

#定义接口数据
paths:
...
  /modules/sys/menu/info/{menuId}:
    get:
      tags:
        - 菜单管理
      summary: 获取菜单信息
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜单ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回菜单信息
          schema:
            $ref: '#/definitions/SysRoleEntity'
  /modules/sys/menu/save:
    post:
      tags:
        - 菜单管理
      summary: 添加菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /modules/sys/menu/update:
    post:
      tags:
        - 菜单管理
      summary: 修改菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /modules/sys/menu/delete:
    post:
      tags:
        - 菜单管理
      summary: 删除菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
...
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      icon:
        description: 菜单图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
...